agenda-flask/requirements.txt
alembic==1.6.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.14
typing-extensions==3.10.0.0
Werkzeug==1.0.1
zipp==3.4.1
